requirements.txt
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
curl_cffi==0.13.0
frozendict==2.4.7
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
gspread==6.2.1
httplib2==0.31.2
idna==3.11
multitasking==0.0.12
numpy==2.0.2
oauthlib==3.3.1
pandas==2.3.3
peewee==3.19.0
platformdirs==4.4.0
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pyparsing==3.3.2
PyYAML==6.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==15.0.1
yfinance==1.1.0
